configure Keycloak

+++ src/main/java/com/example/ms_candidate/CandidatRestAPI.java
@@ -1,5 +1,8 @@
 package com.example.ms_candidate;
 
+import org.keycloak.KeycloakPrincipal;
+import org.keycloak.KeycloakSecurityContext;
+import org.keycloak.adapters.springsecurity.token.KeycloakAuthenticationToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -16,36 +19,54 @@ public class CandidatRestAPI {
     @Autowired
     private CandidatService candidatService;
 
-    @PostMapping("/add")
-    @ResponseStatus
-    public ResponseEntity<Candidat> createCandidat(@RequestBody Candidat c){
-        return new ResponseEntity<>(candidatService.addCandidat(c), HttpStatus.CREATED);
+    @PostMapping("/user")
+    @ResponseStatus(HttpStatus.CREATED)
+    public ResponseEntity<Candidat> createCandidat(@RequestBody Candidat c, KeycloakAuthenticationToken auth) {
+
+        KeycloakPrincipal<KeycloakSecurityContext> principal = (KeycloakPrincipal<KeycloakSecurityContext>) auth.getPrincipal();
+        KeycloakSecurityContext context = principal.getKeycloakSecurityContext();
+        boolean hasUserRole = context.getToken().getRealmAccess().isUserInRole("user");
+
+        if (hasUserRole) {
+            return new ResponseEntity<>(candidatService.addCandidat(c), HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
     }
 
     @GetMapping("/getAll")
     @ResponseStatus
-    public List<Candidat> getAllCandidatsC(){
+    public List<Candidat> getAllCandidatsC() {
         return candidatService.getAllCandidats();
     }
 
     @GetMapping("/getById/{id}")
     @ResponseStatus
-    public Optional<Candidat> getCandidatByIdC(@PathVariable("id")Integer id){
+    public Optional<Candidat> getCandidatByIdC(@PathVariable("id") Integer id) {
         return candidatService.getCandidatById(id);
     }
 
 
-    @PutMapping( value = "/update/{id}",produces = MediaType.APPLICATION_JSON_VALUE)
-    @ResponseStatus
-    public ResponseEntity<Candidat> updateCandidat(@RequestBody Candidat c,@PathVariable("id") Integer id){
-        return new ResponseEntity<>(candidatService.updateCandidat(c,id), HttpStatus.OK);
+    @PutMapping(value = "/update/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
+    @ResponseStatus(HttpStatus.OK)
+    public ResponseEntity<Candidat> updateCandidat(@RequestBody Candidat c, @PathVariable("id") Integer id) {
+        return new ResponseEntity<>(candidatService.updateCandidat(c, id), HttpStatus.OK);
     }
 
-    @DeleteMapping(value = "/delete/{id}")
-    @ResponseStatus
-    public ResponseEntity<String> deleteCandidat(@PathVariable("id") Integer id){
-        return new ResponseEntity<>(candidatService.deleteCandidate(id), HttpStatus.OK);
+    @DeleteMapping(value = "/admin/{id}")
+    @ResponseStatus(HttpStatus.OK)
+    public ResponseEntity<String> deleteCandidat(@PathVariable("id") Integer id, KeycloakAuthenticationToken auth) {
+        KeycloakPrincipal<KeycloakSecurityContext> principal = (KeycloakPrincipal<KeycloakSecurityContext>) auth.getPrincipal();
+        KeycloakSecurityContext context = principal.getKeycloakSecurityContext();
+        boolean hasUserRole = context.getToken().getRealmAccess().isUserInRole("admin");
+        if (hasUserRole) {
+            return new ResponseEntity<>(candidatService.deleteCandidate(id), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
+
 
 
     @PutMapping( value = "/updateLike/{id}",produces = MediaType.APPLICATION_JSON_VALUE)
@@ -59,10 +80,6 @@ public class CandidatRestAPI {
     public ResponseEntity<Candidat> updateCandidatDislike(@PathVariable("id") Integer id){
         return new ResponseEntity<>(candidatService.updateDislike(id), HttpStatus.OK);
     }
-
-
-
-
 
 
 }
